# backend/Api/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "My Title",
    "version": "1.0.0"
  },
  "paths": {
...
    "/Packages/single/{id}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_GetPackage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationPackage"
                }
              }
            }
          }
        }
      }
    },
...
    "/Packages/unfinished/{id}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_GetUnfinishedPackage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresentationPackage"
                }
              }
            }
          }
        }
      }
    },
    "/Packages/new/{type}": {
      "post": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_CreateNewPackage",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/PackageType"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedUnfinishedPackage"
                }
              }
            }
          }
        }
      }
    },
    "/Packages/finish/{id}": {
      "post": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_FinishPackage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FinishedPackage"
                }
              }
            }
          }
        }
      }
    },
    "/Packages/metadata/{id}": {
      "get": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_GetMetadataForPackage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MetadataRecord"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_SaveMetadataForPackage",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "metadataRecords",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MetadataRecord"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/Packages/metapack/{id}/{format}": {
      "post": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_ImportMetadataFile",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "format",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "file",
          "content": {
            "application/octet-stream": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/Packages/properties/{id}": {
      "post": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_SetPackageProperties",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "properties",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageProperties"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/Packages/parameters/{id}": {
      "post": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_SetPackageParameters",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "parameters",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Parameters"
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/Packages/inputs/{id}": {
      "post": {
        "tags": [
          "Packages"
        ],
        "operationId": "Packages_SetPackageInputs",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "inputs",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Action"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "PresentationDevice": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "deviceName": {
            "type": "string",
            "nullable": true
          },
          "deviceShort": {
            "type": "string",
            "nullable": true
          },
          "sensors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/DeviceSensor"
            }
          },
          "packages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PresentationPackage"
            }
          }
        }
      },
      "DeviceSensor": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "$ref": "#/components/schemas/SensorType"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "deviceId": {
            "type": "integer",
            "format": "int32"
          },
          "device": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/PresentationDevice"
              }
            ]
          }
        }
      },
...
      "PresentationPackage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/PackageType"
          },
          "creator": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ToolboxUser"
              }
            ]
          },
          "lastEditedBy": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ToolboxUser"
              }
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "lastEdited": {
            "type": "string",
            "format": "date-time"
          },
          "parametersJson": {
            "type": "string",
            "nullable": true
          },
          "inputsJson": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "$ref": "#/components/schemas/PackageState"
          },
          "pipelineState": {
            "type": "string",
            "format": "byte",
            "nullable": true
          },
          "workDir": {
            "type": "string",
            "nullable": true
          },
          "dataFiles": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/DataFile"
            }
          },
          "metadata": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PackageMetadata"
            }
          },
          "intendedDevices": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PresentationDevice"
            }
          },
          "scripts": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PresentationScript"
            }
          }
        }
      },
      "PackageType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Gallery",
          "Video",
          "Scene",
          "Model"
        ],
        "enum": [
          0,
          1,
          2,
          3
        ]
      },
      "ToolboxUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/IdentityUser"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "firstName": {
                "type": "string",
                "nullable": true
              },
              "lastName": {
                "type": "string",
                "nullable": true
              }
            }
          }
        ]
      },
      "IdentityUser": {
        "allOf": [
          {
            "$ref": "#/components/schemas/IdentityUserOfString"
          },
          {
            "type": "object",
            "additionalProperties": false
          }
        ]
      },
...
      "PackageState": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Unfinished",
          "Processing",
          "Finished"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "DataFile": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "path": {
            "type": "string",
            "nullable": true
          },
          "thumbnailPath": {
            "type": "string",
            "nullable": true
          },
          "previewPath": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PackageMetadata": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PresentationScript": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "packages": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/PresentationPackage"
            }
          },
          "files": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/DataFile"
            }
          },
          "parameters": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ScriptParameter"
            }
          }
        }
      },
      "ScriptParameter": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "key": {
            "type": "string",
            "nullable": true
          },
          "defaultValue": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "CreatedUnfinishedPackage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "FinishedPackage": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "MetadataRecord": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "key": {
            "type": "string",
            "nullable": true
          },
          "value": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "PackageProperties": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "Parameters": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "displayType": {
            "type": "string",
            "nullable": true
          },
          "settings": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Settings"
              }
            ]
          }
        }
      },
      "Settings": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "backgroundColor": {
            "type": "string",
            "nullable": true
          },
          "layout": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Layout"
              }
            ]
          },
          "layoutType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/LayoutType"
              }
            ]
          },
          "padding": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Vector2"
              }
            ]
          },
          "scrollDelay": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "slideAnimationLength": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "cameraAnimation": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/CameraAnimation"
              }
            ]
          },
          "fileName": {
            "type": "string",
            "nullable": true
          },
          "flagInteraction": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/FlagInteraction"
              }
            ]
          },
          "flags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ModelFlag"
            }
          },
          "skybox": {
            "type": "string",
            "nullable": true
          },
          "skyboxTint": {
            "type": "string",
            "nullable": true
          },
          "aspectRatio": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/AspectRatio"
              }
            ]
          },
          "autoStart": {
            "type": "boolean",
            "nullable": true
          },
          "loop": {
            "type": "boolean",
            "nullable": true
          },
          "videoEvents": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/VideoEvent"
            }
          }
        }
      },
      "Layout": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "height": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "horizontalSpacing": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "images": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/GalleryImage"
            }
          },
          "verticalSpacing": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "width": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "spacing": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "visibleImages": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
...
      "LayoutType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Grid",
          "List"
        ],
        "enum": [
          0,
          1
        ]
      },
      "Vector2": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "x": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "y": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
      "CameraAnimation": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "distance": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "height": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "lookAt": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ModelCameraTarget"
              }
            ]
          },
          "origin": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ModelCameraTarget"
              }
            ]
          },
          "revolutionTime": {
            "type": "number",
            "format": "double",
            "nullable": true
          }
        }
      },
...
      "Action": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "mapping": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Mapping"
              }
            ]
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "type": {
            "$ref": "#/components/schemas/TypeEnum"
          }
        }
      },
      "Mapping": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "eventName": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "gestureName": {
            "type": "string",
            "nullable": true
          },
          "condition": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/Condition"
              }
            ]
          },
          "threshold": {
            "type": "string",
            "nullable": true
          },
          "thresholdType": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/ThresholdType"
              }
            ]
          }
        }
      },
...
      "TypeEnum": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "Event",
          "Gesture",
          "GestureDrag",
          "Value",
          "ValueTrigger"
        ],
        "enum": [
          0,
          1,
          2,
          3,
          4
        ]
      }
    }
  }
}